[BE]. Added gd namespace for average color picker realisation, refactoring
